Move fileio from utils to widgets
